refactor(fast_io): re-fold SEC-1 *at* helpers into single at_syscalls module
Consolidates `at_syscalls_metadata.rs` (SEC-1.i) and `at_syscalls_rename.rs`
(SEC-1.j) into the existing `at_syscalls.rs` now that all three SEC-1 *at*
helper PRs are on master. The original module headers committed to a
re-fold once siblings landed; this completes the hygiene follow-up.

- Move metadata helpers (fchmodat / fchownat / utimensat plus their
  via_sandbox_or_fallback adaptors and tests) under a SEC-1.i divider.
- Move rename helpers (renameat plus its via_sandbox_or_fallback adaptor
  and tests) under a SEC-1.j divider. Keep the `RENAME_NOREPLACE` const
  alongside the helpers that consume it.
- Drop the two duplicated `single_component_leaf` private helpers that
  the sibling files carried for in-flight isolation; the original lives
  in `at_syscalls.rs` and is shared by every adaptor.
- Collapse the dir_sandbox/mod.rs re-exports into a single alphabetical
  `pub use at_syscalls::{...}` block.
- Delete `at_syscalls_metadata.rs` and `at_syscalls_rename.rs`.

Zero behaviour changes. All public helper names, signatures, gating, and
the lib.rs re-export surface are preserved.
