+++ crates/fast_io/src/dir_sandbox/mod.rs
@@ -74,27 +74,17 @@ use crate::linux_capabilities::openat2_supported;
 use crate::secure_dir::secure_open_dir;
 
 pub mod at_syscalls;
-#[cfg(unix)]
-pub mod at_syscalls_metadata;
-#[cfg(unix)]
-pub mod at_syscalls_rename;
 
 #[cfg(test)]
 mod tests;
 
 pub use at_syscalls::{
-    AtMetadata, LstatOutcome, UnlinkFlags, fstatat_nofollow, linkat,
-    linkat_via_sandbox_or_fallback, lstat_via_sandbox_or_fallback, mkdirat,
-    mkdirat_via_sandbox_or_fallback, symlinkat, symlinkat_via_sandbox_or_fallback,
-    unlink_via_sandbox_or_fallback, unlinkat,
+    AtMetadata, LstatOutcome, UnlinkFlags, fchmodat, fchmodat_via_sandbox_or_fallback, fchownat,
+    fchownat_via_sandbox_or_fallback, fstatat_nofollow, linkat, linkat_via_sandbox_or_fallback,
+    lstat_via_sandbox_or_fallback, mkdirat, mkdirat_via_sandbox_or_fallback, renameat,
+    renameat_via_sandbox_or_fallback, symlinkat, symlinkat_via_sandbox_or_fallback,
+    unlink_via_sandbox_or_fallback, unlinkat, utimensat, utimensat_via_sandbox_or_fallback,
 };
-#[cfg(unix)]
-pub use at_syscalls_metadata::{
-    fchmodat, fchmodat_via_sandbox_or_fallback, fchownat, fchownat_via_sandbox_or_fallback,
-    utimensat, utimensat_via_sandbox_or_fallback,
-};
-#[cfg(unix)]
-pub use at_syscalls_rename::{renameat, renameat_via_sandbox_or_fallback};
 
 /// Parent-dirfd carrier threaded through the receiver pipeline.
 ///
